Move request ID key to top level tracing package
