+++ api/internal/obfuscation/v1/pkg/interceptor/BUILD.bazel
@@ -10,7 +10,7 @@ go_library(
     visibility = ["//api:__subpackages__"],
     deps = [
         "//api/internal/logkeys",
-        "//api/internal/tracing/v1/pkg/requestid",
+        "//api/internal/tracing",
         "@com_connectrpc_connect//:connect",
     ],
 )
