+++ api/internal/tracing/request_id.go
@@ -1,4 +1,4 @@
-package requestid
+package tracing
 
 // RequestIDKey represents the string used as the key to access the request ID value from the request metadata map.
 const RequestIDKey = "request-id"
